pyproject.toml
[project]
name = "st"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "floggit>=0.0.19",
    "google-cloud-storage>=3.3.0",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "streamlit>=1.48.0",
    "streamlit-agraph>=0.0.45",
    "streamlit-extras>=0.7.6",
    "streamlit-mermaid>=0.3.0",
]
